Fix layering violation in WebCore's PathUtilities.h https://bugs.webkit.org/show_bug.cgi?id=304678
Reviewed by Alan Baradlay.

Fixes layering violation in WebCore's PathUtilities.h by moving
PathUtilities::pathWithShrinkWrappedRectsForOutline(...) to live
on OutlinePainter.

The issue was that `pathWithShrinkWrappedRectsForOutline` used
non-platform layer types `Style::BorderRadius` and `WritingMode`,
so, the function needed to live above the platform layer.

To make this possible, the `FloatPointGraph` that used to live
in PathUtilities.cpp was moved into its own files, along with
the `polygonsForRect` utility function.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/FloatPointGraph.cpp: Added.
* Source/WebCore/platform/graphics/FloatPointGraph.h: Added.
* Source/WebCore/platform/graphics/PathUtilities.cpp:
* Source/WebCore/platform/graphics/PathUtilities.h:
* Source/WebCore/rendering/OutlinePainter.cpp:
* Source/WebCore/rendering/OutlinePainter.h:

Canonical link: https://commits.webkit.org/304937@main
