+++ Source/WebCore/rendering/OutlinePainter.h
@@ -27,9 +27,14 @@
 
 namespace WebCore {
 
+class Path;
 class RenderInline;
 class RenderListBox;
 struct PaintInfo;
+
+namespace Style {
+struct BorderRadius;
+}
 
 class OutlinePainter {
 public:
@@ -50,6 +55,8 @@ private:
     static bool collectFocusRingRectsForBlock(const RenderBlock&, Vector<LayoutRect>&, const LayoutPoint&, const RenderLayerModelObject*);
     static void collectFocusRingRectsForInlineChildren(const RenderBlockFlow&, Vector<LayoutRect>&, const LayoutPoint&, const RenderLayerModelObject*);
     static void collectFocusRingRectsForChildBox(const RenderBox&, Vector<LayoutRect>&, const LayoutPoint&, const RenderLayerModelObject*);
+
+    static Path pathWithShrinkWrappedRects(const Vector<FloatRect>&, const Style::BorderRadius&, float outlineOffset, WritingMode, float deviceScaleFactor);
 
     void addPDFURLAnnotationForLink(const RenderElement&, const LayoutPoint& paintOffset) const;
